requirements.txt
Flask>=2.0.0
python-dotenv>=0.21.0
python-telegram-bot>=13.7
requests>=2.28.0
gunicorn>=20.1.0
